separated the the removal function and added tests

+++ sdk/src/main/java/ly/count/android/sdk/CountlyStore.java
@@ -85,7 +85,7 @@ public class CountlyStore implements StorageProvider, EventQueueProvider {
     ConfigurationProvider configurationProvider;
 
     int maxRequestQueueSize = 1000;
-    static int dropAgeHours = 0;
+    int dropAgeHours = 0;
 
     //explicit storage fields
     boolean explicitStorageModeEnabled;
@@ -115,6 +115,15 @@ public class CountlyStore implements StorageProvider, EventQueueProvider {
 
     public void setLimits(final int maxRequestQueueSize) {
         this.maxRequestQueueSize = maxRequestQueueSize;
+    }
+
+    /**
+     * For testing purposes. Sets the dropAgeHours
+     *
+     * @param dropAgeHours
+     */
+    public void setRequestAgeLimit(final int dropAgeHours) {
+        this.dropAgeHours = dropAgeHours;
     }
 
     public void setConfigurationProvider(ConfigurationProvider configurationProvider) {
@@ -365,7 +374,11 @@ public class CountlyStore implements StorageProvider, EventQueueProvider {
                 //reached the limit, start deleting oldest requests
                 L.w("[CountlyStore] Store reached it's limit, deleting oldest request(s)");
 
-                deleteOldestRequest();
+                // TODO: too much I/O?
+                int removedRequests = checkAndRemoveTooOldRequests(); // remove too old requests
+                if (removedRequests == 0) { // remove oldest if nothing is too old
+                    deleteOldestRequest();
+                }
                 addRequest(requestStr, writeInSync);
             }
         } else {
@@ -373,33 +386,59 @@ public class CountlyStore implements StorageProvider, EventQueueProvider {
         }
     }
 
+    /**
+     * Removes the oldest request:
+     * 1. Gets the current requests from storage
+     * 2. Copies it to a List
+     * 3. Removes the first entry
+     * 4. Saves the list to the storage
+     */
     synchronized void deleteOldestRequest() {
         final List<String> connections = new ArrayList<>(Arrays.asList(getRequests()));
 
-        int removedReqs = 0;
+        L.i("[CountlyStore] deleteOldestRequest, Will remove the oldest request");
+        connections.remove(0);
+
+        storageWriteRequestQueue(Utils.joinCountlyStore(connections, DELIMITER), false);
+    }
+
+    /**
+     * Removes too old requests from the queue:
+     * 1. Checks if there is a request age limit set
+     * - If not returns 0
+     * - If there is:
+     * 1. Gets the current requests from storage
+     * 2. Copies it to a List
+     * 3. Iterates over it and removes the old ones
+     * 4. Saves the list to the storage
+     * 5. Returns the number of removed requests
+     */
+    synchronized int checkAndRemoveTooOldRequests() {
+        int removedRequests = 0;
+
         // if there is a request age limit set, check the whole queue for older requests
         if (dropAgeHours > 0) {
-            L.i("[CountlyStore] deleteOldestRequest, will remove outdated requests from the queue");
-            Iterator<String> iterator = connections.iterator();
+            final List<String> connectionsList = new ArrayList<>(Arrays.asList(getRequests()));
+            L.i("[CountlyStore] checkAndRemoveTooOldRequests, will remove outdated requests from the queue");
+            Iterator<String> iterator = connectionsList.iterator();
             while (iterator.hasNext()) {
                 String request = iterator.next();
 
                 // check if the request is too old, and remove it from the list
                 if (Utils.isRequestTooOld(request, dropAgeHours, "[CountlyStore]", L)) {
-                    L.v("[CountlyStore] deleteOldestRequest, removing:" + request);
+                    L.v("[CountlyStore] checkAndRemoveTooOldRequests, removing:" + request);
                     iterator.remove();
-                    removedReqs++;
+                    removedRequests++;
                 }
+            }
+
+            // save the new request queue
+            if (removedRequests > 0) {
+                storageWriteRequestQueue(Utils.joinCountlyStore(connectionsList, DELIMITER), false);
             }
         }
 
-        // if no age limit or no old reqs in queue. just remove single request
-        if (removedReqs == 0) {
-            L.i("[CountlyStore] deleteOldestRequest, no request age limit or no too old requests. Will remove the oldest request only");
-            connections.remove(0);
-        }
-
-        storageWriteRequestQueue(Utils.joinCountlyStore(connections, DELIMITER), false);
+        return removedRequests;
     }
 
     /**
